:package: create authenticate routes

+++ source/routes/memories.ts
@@ -3,8 +3,15 @@ import { z } from 'zod'
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -18,18 +25,23 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
     const { id } = paramsSchema.parse(request.params)
-
     const memory = prisma.memory.findUniqueOrThrow({
       where: {
         id,
       },
     })
 
+    if (
+      !(await memory).isPublic &&
+      (await memory).userId !== request.user.sub
+    ) {
+      return reply.status(401).send()
+    }
     return memory
   })
 
@@ -40,19 +52,18 @@ export async function memoriesRoutes(app: FastifyInstance) {
       isPublic: z.coerce.boolean().default(false),
     })
     const { content, isPublic, coverUrl } = bodyShema.parse(request.body)
-
     const memory = await prisma.memory.create({
       data: {
         content,
         coverUrl,
         isPublic,
-        userId: '6370296a-ca76-45bf-9c42-c8291b9036dc',
+        userId: request.user.sub,
       },
     })
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -66,7 +77,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
     const { content, isPublic, coverUrl } = bodyShema.parse(request.body)
 
-    const updateMemory = await prisma.memory.update({
+    let updateMemory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (updateMemory.userId !== request.user.sub) {
+      return reply.status(401).send
+    }
+
+    updateMemory = await prisma.memory.update({
       where: {
         id,
       },
@@ -79,12 +100,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return updateMemory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send
+    }
 
     await prisma.memory.delete({
       where: {
